Add eirini-daemonset to helm chart
instead of creating it in a post-deployment job. This is neccessary to
eventually run the post-deployment job without cluster-admin privileges.

It also makes it much easier to modify the daemonset template before
deployment to adjust to nostandard requirements.

The daemonset currently hard-codes the loggregator/agent:latest image
name from DockerHub. This image doesn't use Configgin, so it is not
possible to use KUBERNETES_CLUSTER_DOMAIN, and the value has been
hard-coded to cluster.local.